Consolidate XML serialization helpers
Generate the XML root wrapper in the existing XML helper file alongside additional-properties support.

Copilot-Session: a2fed377-7b26-45fc-b3aa-8b6d6df8b46a
